SLF4J with Simple provider
---
https://www.slf4j.org/manual.html#swapping:~:text=Binding/provider%20for%20Simple%20implementation

Binding/provider for Simple implementation, which outputs all events to System.err. Only messages of level INFO and higher are printed. This binding may be useful in the context of small applications.

---

https://www.slf4j.org/manual.html#swapping:~:text=The%20warning%20will%20disappear

The warning will disappear as soon as you add a provider to your class path. Assuming you add slf4j-simple-2.0.9.jar so that your class path contains:
  - slf4j-api-2.0.9.jar
  - slf4j-simple-2.0.9.jar